Move Rails SLI initialization to boot path
This was done lazily because it comes with several
seconds of delay, since we need to traverse all
web and API endpoints.

However, this caused strong coupling between the
WebExporter and Rails, the former we are looking
to first extract and then retire.

Rails boot performance is a known issue, but adding
the delay does not dramatically increase the
existing timing at least during local measurements.

A pending optimization could be to only compile
Grape routes in the api fleet, but we have no
mechanism currently for Rails to know which
workload it serves.