Update to latest webpack-dev-server
This fixes a bunch of dependency audit problems. Luckily the new
webpack-dev-server is compatible with webpack v4 and webpack v5.

Updating this separately from the rest of webpack seems like a step in
the right direction. So far there are no problems with the webpack dev
server locally and HMR and all these things seem to work perfectly.