Fix Geo checksummable check failing when file is nil
In specific cases, such as the LFS download flow,
we would first create a model before updating it to contain
a file as well, which means it's possible the model's file
doesn't exist.

We have an after_commit hook that attempts to verify if
a record's file is checksummable to run verification on,
however this would have failed if the file was nil before this.

Changelog: fixed
EE: true