Backfill SSL verification for integrations with known-good hostnames
Some integrations were hard-coded to disable SSL verification, and we
added this setting in a security fix and enabled it by default for
new configurations.

For existing configurations this was handled in the application code,
but now that the security fix is released we can update the settings in
the DB, and then in the next release we can remove our temporary code
in the app again.

Changelog: other