Allow testing of disabled hooks
Ignore executable status if the webhook is executed directly, as it
is when testing a hook.

This is essential for dealing with disabled hooks. If they cannot be
run, then they cannot be re-enabled. For this reason we also run the
log-execution (which handles re-enablement) inline during tests.

Changelog: fixed