Fix namespace statistics migration failure
https://gitlab.com/gitlab-org/gitlab/-/merge_requests/79657 modified
`NamespaceStatistics` to use a new database column
(`dependency_proxy_size`) in the `before_save` callback. However,
migration tests that roll back to an older schema value previously
would fail since the column did not exist.

To mitigate this issue, we now check for the existence of the column
before attempting to use it.