Use cached column list for members union query
When doing a UNION query against a table that has a newly
added column, if the Rails code has not yet been restarted,
there can be a mismatch of columns, resulting in the UNION
query failing. This can happen during a rolling deploy across
a fleet of servers.

This fixes one such instance.

Changelog: fixed