Add deployment approval UI MVC
Sets up the ability for users to approve and reject deployments from
within the environments page.

Current information shown is how many approvals are needed and how many
are done, as well as who has approved and when.

A rejection short cuts to killing the deployment.

This is currently only in the new environment UI, as the old one is
going away.

Changelog: added
EE: true