Do not use transactional tests for this feature spec
The setup for this spec relies on multiple threads,
but it hangs with transactional tests + decomposed databases.

We could re-initialize database load balancing to re-establish a correct
connection to prevent the hang, but now we cannot see the data we
previously created. So the solution is to not use transactional tests,
and rely on database cleaner's deletion strategy