Add ref param to project level ci/lint
Refactor issue move to not require group_id param

This simplifies the code so that we allow passing in move_before_id /
move_after_id as long as they are part of the same project / namespace
hierarchy.

This also fixes the reorder API so that it now accepts IDs from other
projects within the same hierarchy.

Changelog: fixed