Improve UI message for unfinished pipelines
This message should be revised to meet our standards - include
current status, explanation, next step.

Update tests and translation file

Update the .pot file for translation, and update a test that checks
the output of the file we modified in this MR.

Update messages based on UX feedback

We've decided to NOT change one of the messages, and revising the
other. Update tests and translation files to match.

Update test and snapshot

The text was just a smidge off from what the tests were expecting.
Move the period inside the link, rather than outside of it, to strip
the extra space I'm seeing.

Update translations file

Update translations file to reflect changes to strings.

Changelog: changed