Allow guests to set issue position on create
Update permission should not be required since we're not really updating
the sibling issues. Their relative position values may change if there
is not enough space and we need to shift some issues, but these numbers
are purely internal and does not change the position of the issue
relative to the existing issues.