Refactor upload package
In https://gitlab.com/gitlab-org/gitlab/-/issues/351657, the
goal is to document and refactor workhorse code related to
object storage upload. This is to improve readability and
help new contributors.

This include the following changes:
- Moved `UploadArtifacts` to upload package to consolidate all
  uploaders in a single package. Renamed it to `Artifacts`
  to be consistent with other uploaders.
- Unexported `InterceptMultipartFiles` to reduce the number of
  exported functions in upload package.

These upload middlewares are not yet mentioned in developer docs.
A separate change will be made to add information about these
uploaders and which one to use.