Global Search - Optimize SearchContext
This change adds a helper method for the
new Header Search.  This is realted to the
SearchContext Class sharing too much info
with the frontend.

Instead we now explicitly extract the
data we need before sharing it with the
frontend.  This cleans up the DOM quite
a bit and helps guard from leaking
any data.

Changelog: changed