Restart Action Cable server when Redis disconnects
Prevent Puma from crashing when there are connection errors to Redis.
This change catches the error and restarts the Action Cable server when
there are connection errors.

We need to restart Action Cable and disconnect all clients because once
we reconnect to Redis, all the previous SUBSCRIBE commands we made are
no longer applicable.

Changelog: fixed