Implement feature flag to rollout decomposed primary CI database
Per our rollout plan https://gitlab.com/groups/gitlab-org/-/epics/6160
of separating the CI database from the main database the "phase 4" will
involve gradually changing a percentage of "read-write" (primary)
traffic to go through a new connection that still points to the same
database.

Previously we added the `GITLAB_LOAD_BALANCING_REUSE_PRIMARY_ci`
environment variable to allow us to share the same primary connection
when there is a separate `ci` entry in `config/database.yml`. But we
need to a way to override this environment variable so this MR
introduces a new feature flag `force_no_sharing_primary_model` which
overrides the "reuse" connection and forces it to not reuse the
connection.