Rename `subject` to `__subject__` in delegated presenters
We want to encourage developers to use an explicit named getter instead
using `presents ..., as: <getter>`, since the `subject` name clashes
with some model attributes (like in the case of labels), and can be
confusing in general.

The `__subject__` getter can still be used for generic "internal" use,
for example when we need to unwrap a presenter in GraphQL's
`BaseResolver`.

This also removes the need for some `delegator_override` calls in
presenters.