Fairly large redo of docs
This MR moves all of the GitOps and CI/CD topics
into their own pages. It moves topics out of the Install
page and onto the Configuration/Working with page.
It uses lowercase for the agent.
The intention with these changes is to make the information
easier to find and build on in the future