Add rate limiter for PullRequest importer
* Contributes to https://gitlab.com/gitlab-org/gitlab/-/issues/345922
* Feature flag: https://gitlab.com/gitlab-org/gitlab/-/issues/353217

**Problem**

We don't have a limitation for a number of PullRequest importer
jobs. The parallel importer process can generate thousands of them. It
puts pressure on other services.

**Solution**

Limit the number of created jobs by spreading the load for a larger
period of time.

New settings `max_concurrency` and `backoff` allow to setup the
desired size and delay between batches.