Cache namespace first Auto DevOps config
Retrieving a namespace's first Auto DevOps config can be an
expensive task. The value cascades so the method queries up the
hierarchy to find the first non-nil value.

This change caches that result to reduce the frequency of the N+1
necessary to get these results. Since the setting should rarely
change, it's helpful to cache the result and selectively expire
the cache.

Changelog: fixed