Add sections definition to integration fields
We are redesigning the page to be split into sections
to make it easier for users to configure an integration.

This change adds the sections definition (which fields
belong to a section, section title, and description) to
the backend model and passes it to the frontend.
It's added this way to keep all integration related data
in the backend.