Ensure cleanup job artifacts task does not include pipeline artifacts
Pipeline artifacts share the artifacts storage config and generate
a similar path that instead of the date (for job artifacts) would be
"pipelines", which also matches this here.

Because the cleanup **job** artifacts rake task is looking for job
artifacts only in the database, this will not find pipeline artifacts
and think they are orphaned, deleting them from disk.

Changelog: fixed