Make GitlabSlackApplication available to test env
fdf31bb2 made the Integrations::GitlabSlackApplication integration
available to development in addition to GitLab.com.

When a user now generates the GraphQL docs with
`rake gitlab:graphql:compile_docs` locally, the new integration appears
as a value in the enum `ServiceType`
https://gitlab.com/gitlab-org/gitlab/-/blob/b8c957ad50e4ffafe66b6869fc42a936e8aab5ac/app/graphql/types/projects/service_type_enum.rb
however because the integration is not available to the test environment
the `graphql-verify` job will fail because
`rake gitlab:graphql:validate` will check the commited changes against
its own version of the GraphQL docs, where the integration value will be
absent from the enum.

This change makes the integration available to test so the integrations
are consistent between a developer's local environment and the
`graphql-verify` job.

https://gitlab.com/gitlab-org/gitlab/-/merge_requests/79304/diffs#note_822981597