Add FF to toggle cert-based clusters features
We're deprecating cert-based clusters and all of its related features
will be removed. So we've decided to use feature flags to hide the
features in the major milestone instead of right away removing them.
This MR introduces the first usage of this FF, but other MRs will come
to put this FF in other places.

At this time, we're hiding everything that makes use of
deployment_platform. The most impactful thing is not creating rollout
jobs anymore during Auto DevOps pipelines. So we're also adding tests
for this.