Fix TypeError from Tags::Reference
Some missing references were raising a TypeError when using `dig` to
extract the referenced object from a reference, in cases where the
reference included too many symbols and we ended up attempting to
extract a value from an array using a symbol.

This commit fixes the issue by rescuing from TypeError and properly
returning a message stating that the given reference cannot be found in
the given configuration.

Changelog: fixed