Implement Gitlab mirror scheduling tracker
This tracker is used to track the number of projects are being
scheduled. This number is used in UpdateAllMirrorsWorker to control its
rescheduling frequency. This number is also a replacement for
UpdateAllMirrorsWorker#queue_size, in an initiative to remove the
dependency to Sidekiq queues.

We made two attempts to resolve this issue, but not successfully:
- https://gitlab.com/gitlab-org/gitlab/-/merge_requests/80711
- https://gitlab.com/gitlab-org/gitlab/-/merge_requests/79097

Changelog: other