Remove workers workaround in Pipeline Editor
This fixes a problem for GitLab instances set to use a CDN, via the
`GITLAB_CDN_HOST` environment variable. They were unable to fetch the CI
schema as the path ended up being a concatentation of two absolute URLs.

If the schema path is absolute, it's used as-is. Otherwise, it's made
absolute using the GitLab instance as the base.