Consolidate upload preparers
In https://gitlab.com/gitlab-org/gitlab/-/issues/351657, the
goal is to document and refactor workhorse code related to
object storage upload. This is to improve readability and
help new contributors.

All preparer types are consolidated in the `upload` package so
they are easier to find and show that when a new type is needed,
this is where it can be added.

Currently, preparers are not mentioned in developer docs. A
separate change will be created for it to guide contributors
whether they need to add a new preparer or use an existing one.