Adds a notification for a future dated license
When an instance has no currently active subscription, but has a future
dated subscription — we display a notification to ensure them that they
don't need to do anything else to activate it. And the Subscription
History table is now shown whether there is active subscription or not.
Placing it above "Free trial" and "Subscription" cards if there's a
future dated subscription, and under them if not.

Changelog: changed
EE: true