Introduce a field DSL for integrations
This aims to improve the developer experience for integrations.

Currently for every accessible field, we need to define prop
accessors and field definitions. This change introduces a couple of
class-level DSL methods to make this clearer, less error prone and less
duplicative, with sensible defaults in some cases.

A Field class is introduced to make transformations of this data easier.