Introduce concern for resetting secret fields
We want to force users to re-enter secret fields when changing
certain other fields (such as URLs), which can unintentionally expose
the secret fields.

This adds a new concern to handle this logic, and sets it up in the
Jira integration. In future MRs we also want to reuse this concern
for other integrations which currently duplicate this logic.

Exposing fields can be marked with the `exposing_secrets:` flag in
the Field DSL.

The Jira integration also had some custom logic where the `url` field
was only considered if the `api_url` was blank, but to generalize this
we're now resetting the secrets if either of the URL fields has changed.