Change Edit to Open in Blob view to maintain consistency
The text for Edit in Web IDE and Edit in Gitpod in Blob file view were not consistent with Merge Request actions. For consistency, change Edit in Web IDE to Open in Web IDE and Edit in Gitpod to Open in Gitpod. Add new label Open in Gitpod. Fix the test case by changing the button text to Open in Web IDE.

Changelog: fixed