Refactor repo deletion existence checking
In this commit we're changing the way we rely on the
repo existence in order to remove the repo. It may happen
that we encounter a race condition between the moment
we check the repo existence and the moment we delete it.

In this commit we catch the RPC exception when the repo
does not exist in order to continue the operation.

Changelog: fixed