Do not disable snippet form submit button to enforce validation
Pajamas guidelines states that all form fields not marked as (optional)
should be required and that submit buttons should not be disabled when
there are validation errors in the form.

This update the snippet form implementation to follow the Pajamas
guidelines. It updates the form to:

1. Display "This field is required." error message on the Title and
Files fields when left blank and submit button is clicked
2. Removes the implementation that disables the submit button when Title
and Files fields are left blank

Changelog: changed