Fix sending BlobContent query with empty variables
In the pipeline editor, we make an initial query to get
the BlobContent, and then we parse this content through
the configData query. However, to make that initial query,
we rely on 2 client side queries for the branchName and the
ref. We have to wait for both of these query to be resolved
before we can fetch the blob, otherwise we send malformed
requests.

Changelog: fixed