Remove merge_request_eager_fetch_ref feature flag
https://gitlab.com/gitlab-org/gitlab/-/merge_requests/80876 fixed a
problem where the Gitaly `FetchSourceBranch` RPC would cause the
database to idle in transaction. This happened because the call was
made in an `after_save` hook, which wraps everything in a transaction.
On GitLab.com, the 15-second idle in transaction timeout would close
the connection, but this was only apparent to the application when it
attempted to write a merge request metrics entry for a non-existent
merge request, resulting in a foreign key violation.

With the feature flag enabled, the `FetchSourceBranch` occurs outside
this hook. Since enabling this feature flag, we no longer see this
idle in transaction problem.

Relates to:

1. https://gitlab.com/gitlab-org/gitlab/-/issues/336657
2. https://gitlab.com/gitlab-org/gitlab/-/issues/353044