Update maximum allowable lifetime message for PAT
Admins can set a maximum allowable lifetime for personal access
tokens. When this setting is active a message is displayed to users
so they know what is the maximum date that can be set for expiration.
This commit updates the copy and styling of this message.

Changelog: changed
EE: true