UI Bug Fix: Expand the early checks when using /merge quick action
At the moment we do a check of the `mergable_state?` method to see
if it can be merged, and later when we try to merge, the
`mergeable?` method is called. But sometimes this means we return
that the merge request was merged, when it actually isn't.

Issue: gitlab.com/gitlab-org/gitlab/-/issues/354453
MR: gitlab.com/gitlab-org/gitlab/-/merge_requests/82170

Changelog: fixed
EE: true