Display license ID retrieved from CustomersDot
Prior to this changes, the displayed license ID was the one from
GitLab's license database table instead of the one retrieved from
CustomersDot. While the right logic was supposed to be used, the
global_id logic used the license's id instead of the license_id method
as used in the graphql resolvers.
This change will fix the issue with the global id logic by overriding
the id method in the resolver with a custom global id logic that uses
the license_id method.

Changelog: changed
EE: true