Remove constraints from sprints table
As part of the plan to make title an optional attribute for Iteration,
the non-null constraint is removed along with the uniqueness constraints
from the database while the application-level checks are preserved.

We are including a down migration to set any null titles to their
primary keys (ids) because we want to start accepting null titles
behind a feature flag in the same release.

Changelog: changed