Disable multi-project viz for free users
There was a bug where we would allow free tier
users to see multi project pipelines but this is
a premium feature. We now disable the expand
button for free tier users and show a popover
explaining why this is disabled and adding tests
to ensure it works as expected and no regression
occurs again.

Changelog: changed