Allow gitaly create_repository to take a default branch argument
Currently when rails wants to ensure a default branch name, it will call
the Gitaly WriteRef RPC to manually change it. If this call happens in a
racy way eg: it gets called when another RPC is operating on the same
repository, it will fail. Rather than do this, the Gitaly
CreateRepository RPC has a default_branch parameter.

Allow rails callers to set this value. If it is not provided, then the
default branch in Gitaly will be used, which is main.