Add support stackprof in GitLab profiler
Stackprof runs faster and makes it easy to generate a quick profile
and flamegraph for a given endpoint. This commit adds support for
Stackprof in `Gitlab::Profiler.profile` via the `sampling_mode: true`
parameter.

By default, the sampling data will be written to a temporary file, but
`profiler_options` can be used to specify the destination and other
options.

Changelog: added