Instructions to run pending background migrations
This change adds commands to re-run all migrations found marked as
pending and remaining stuck in such a state

It expands on an existing area of troubleshooting
background migrations while specifically
targeting the pending state
situation instead of migration
jobs that are scheduled but remain un-executed.

It also adds an upgrade note to version 13.6 section where a background
migration was introduced with a bug in it that can
cause it to show up as pending in all releases after it. This was only
recently noticed after we introduced new commands
that check for pending background migration jobs
as part of checking background migrations prior to upgrades.

A fix for the 13.6 bug should be included in GitLab 14.9 onwards via
https://gitlab.com/gitlab-org/gitlab/-/merge_requests/81879
and following the added instructions should help
clear the migration.