Do not raise if selected item is not found
As noticed in
https://gitlab.com/gitlab-org/gitlab/-/merge_requests/80681#note_862019211
and
https://gitlab.com/gitlab-org/gitlab/-/merge_requests/81370#note_854661919,
the result of raising if the selected item cannot be found is a 500
error. Considering this can be triggered by user behaviour (i.e.,
changing the URL) the only sensible way to handle this is to select the
first item, as was already the case for `selected: nil`.

Addresses https://gitlab.com/gitlab-org/gitlab/-/issues/354424.