Include merge-requests in RESERVED_REFS_NAMES
When importing a project currently we delete the merge-request refs,
but for an in-progress MR we want to keep these.

Looking at the history of the reserved ref names, in the initial
commit to the cleanup service, 5f811894, merge-requests was included
in the list, but in 081e2fce we transferred the list to the repository
model and inadvertently duplicated environments and left out
merge-requests.  We fixed the duplicate environment entry with
ba6d3fcb, but never re-added merge-requests.

This commit re-adds merge-requests to the reserved reference list.
No additional tests added because the existing tests already
validate that all reference types in RESERVED_REFS_NAMES are
not removed.

Changelog: fixed