Add frontend routing to Geo GraphQL specific sites
This adds replicable endpoints for each Geo site, to
allow the primary to request this data from the tracking DB
directly on any secondary site.

This is necessary for Geo proxying with unified URLs, as
the web traffic will all be proxied to the primary, however
we don't have projects/designs implemented in GraphQL so these
don't yet work.

Changelog: added
EE: true