Add frontend for connection section
When the integrationFormSections feature flag is
enabled and `sections` is provided from the backend,
we now add a separate "Connection details" section.

For now, this section will include ActiveCheckbox and
some of the DynamicFields which have been filtered
by the backend.

We use dynamic imports here since not all integrations
have this sections defined yet and there will be more sections
that are integration specific (for example, Jira).