Update Auto DevOps docker versions to 20.x
Mainly a routine update. Only noteworthy chagne is that due to changes
in Docker 20.x, we must now pass the `--tls=false` flag for
compatibility with runners that do not have configured a volume for
sharing the the TLS certificate between the docker-in-docker service
container and the job container.

See https://gitlab.com/gitlab-org/gitlab/-/issues/328208

Changelog: changed