Fix filters presence check to take value into consideration
This fixes a bug in the issues/MRs dashboard filter check which
allows filtering by empty value. It can result to SQL statement
timeout.

The fix is to check if the value of the filter is present. If not,
the dashboard will show the "Please select at least one filter to
see results" message.

Changelog: fixed