Add deployment approval comment field
A comment field allows users to describe why a deployment is approved or
rejected.

It shows a count of remaining characters. The count turns orange when 30
or less are left, and turns red once the comment is over the limit.

This counter should be added to GitLab UI (and made pajamas compliant),
but is good here for now.

Changelog: added
EE: true