New Group Runners administration view
This change enables the new group runners view via the CI/CD menu in
the left navigation bar.

Group owners can now more easily administer and manage the Runners
associated with the namespace. The new view includes the summary
count of the number of online, offline, and stale runners associated
with the group and sub-groups.

Changelog: added