Don't allow filtering by `in` alone on issue/MR dashboard
When `in` param is set while no `search` param is set, the issues
and MRs finder won't use the `in` param. It'll instead skip the
filter check and can result to SQL statement timeout.

The fix is to check that if `in` param is set but no search, show
the "Please select at least one filter to see results" message.

Changelog: fixed