Align `The` usage in API field docs
I've recently learned during another Review for one of my MRs that a
leading `The` for field desriptions is not necessary
(see https://gitlab.com/gitlab-org/gitlab/-/merge_requests/82467#note_869500543).

Therefore, I have omitted it for the `GET` APIs and this change now
aligns the documentation for the Group and Project Access Tokens.