Add fixture guard spec for SQL metric queries
This commit adds a guard spec that compares SQL metric queries to a
fixture file in order to catch any arbitrary changes to the queries.

The fixture file can be regenerated using a rake task.

Timestamps are freezed for and a stable set of Devise omniauth providers
is used to make queries that rely on them deterministic and not change
every time the fixture file is regenerated.