Consistently load backup manifest
Previously it was not clear whether one should use `settings` or
`backup_information`. This resulted in both being used inconsistently.
Instead of silently loading backup information, we explicitly load or
build the information depending.

Now that we can rely on backup information being consistent, this means
we can now start using it to detect when tasks are skipped instead of
adhoc reading from ENV.