gl-emoji: Move fallback image rendering to frontend
In 2d5b14e3f1b (Add CustomEmoji Banzai Filter, 2020-11-07) a first
iteration of rendering custom emoji in Banzai was added. That change
inserted <img> tags on the backend and made sure the frontend did not
clean them out.

This change modifies that and passes the fallback image source to the
frontend so it can insert <img> tags in frontend code.